Extract and export the `map` serializer
