+++ packages/umi-serializer-beet/src/index.ts
@@ -2,6 +2,7 @@ export * from './array';
 export * from './createBeetSerializer';
 export * from './bytes';
 export * from './errors';
+export * from './map';
 export * from './tuple';
 export * from './plugin';
 export * from './pubkey';
