+++ packages/umi-serializer-data-view/src/index.ts
@@ -2,6 +2,7 @@ export * from './array';
 export * from './createDataViewSerializer';
 export * from './bytes';
 export * from './errors';
+export * from './map';
 export * from './tuple';
 export * from './plugin';
 export * from './pubkey';
