+++ packages/umi-serializer-data-view/test/map.test.ts
@@ -1,10 +1,11 @@
 import test from 'ava';
 import { Endian } from '@metaplex-foundation/umi';
+import { map } from '../src/map';
 import { createDataViewSerializer } from '../src';
 import { s, d } from './_helpers';
 
 test('prefixed (de)serialization', (t) => {
-  const { map, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, map(u8(), u8()), new Map(), '00000000'); // 4-bytes prefix.
@@ -34,7 +35,7 @@ test('prefixed (de)serialization', (t) => {
 });
 
 test('fixed (de)serialization', (t) => {
-  const { map, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
 
   // Empty.
   s(t, map(u8(), u8(), { size: 0 }), new Map(), '');
@@ -73,7 +74,7 @@ test('fixed (de)serialization', (t) => {
 });
 
 test('remainder (de)serialization', (t) => {
-  const { map, u8, string, u64 } = createDataViewSerializer();
+  const { u8, string, u64 } = createDataViewSerializer();
   const remainder = { size: 'remainder' } as const;
 
   // Empty.
@@ -111,7 +112,7 @@ test('remainder (de)serialization', (t) => {
 });
 
 test('description', (t) => {
-  const { map, u8, u16, string } = createDataViewSerializer();
+  const { u8, u16, string } = createDataViewSerializer();
 
   // Size.
   t.is(map(u8(), u8(), { size: 42 }).description, 'map(u8, u8; 42)');
@@ -138,7 +139,7 @@ test('description', (t) => {
 });
 
 test('sizes', (t) => {
-  const { map, u8, u32, string } = createDataViewSerializer();
+  const { u8, u32, string } = createDataViewSerializer();
   t.is(map(u8(), u8()).fixedSize, null);
   t.is(map(u8(), u8()).maxSize, null);
   t.is(map(u8(), u8(), { size: u8() }).fixedSize, null);
